requirements.txt
Flask==3.0.0
imagehash==4.3.1
Pillow==10.0.0
